Reformat code structure
